pyproject.toml
[project]
name = "forecast-presentation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "altair>=5.5.0",
    "datasets>=3.6.0",
    "epiweeks>=2.3.0",
    "marimo>=0.13.15",
    "matplotlib>=3.10.3",
    "mosqlient>=1.9.3",
    "pyarrow>=20.0.0",
    "python-dotenv>=1.1.0",
    "pywavelets>=1.8.0",
    "ruff>=0.11.13",
    "vega>=4.1.0",
]
